% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accident_package.R
\name{getPlaceDataAsList}
\alias{getPlaceDataAsList}
\title{getPlaceDataAsList}
\usage{
getPlaceDataAsList(city, county, state)
}
\arguments{
\item{city}{city name}

\item{county}{county name}

\item{state}{state name}
}
\value{

}
\description{
getPlaceDataAsList
}
\examples{
getPlaceDataAsList("Chicago", "Cook", "IL")
getPlaceDataAsList("city", "county", "state")
}
