% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accident_package.R
\name{getPlaceDataAsPoints}
\alias{getPlaceDataAsPoints}
\title{getPlaceDataAsPoints}
\usage{
getPlaceDataAsPoints(city, county, state)
}
\arguments{
\item{city}{city name}

\item{county}{county name}

\item{state}{state name}
}
\value{

}
\description{
getPlaceDataAsPoints
}
\examples{
getPlaceDataAsPoints("Chicago", "Cook", "IL")
getPlaceDataAsPoints("city", "county", "state")
}
